fix pthread_create race
Fix a possible segfault caused by a race with the thread variable
returned by pthread_create and the use of that variable in the
thread. SUSv3 explicitly notes that the implementation need not
initialize the buffer pointed to by @thread@ before the new thread
starts executing; that is, the new thread may start running before
pthread_create() returns to its caller. Fix this race by moving the
mutex lock/unlock earlier such that we guarantee that the thread
variable rt->t1.thread is valid before it is used.