sess_datapath.sh: be more tolerant of error exit codes
When testing on Alpine Linux, an ip netns exec command that is
expected to fail because the given namespace does not exist returns an
exit code of 255 instead of 1. Modify the script to check that the
exit code is any non-zero value since we only care that the command
returned non-zero.