fix test failures when kernel delays l2tp tunnel removal
When userspace closes a tunnel socket, the kernel closes the
associated L2TP tunnel along with all of its sessions (if any). The
actual tunnel context is destroyed by the kernel asynchronously. It is
therefore possible that an attempt to create a tunnel with the same ID
after a tunnel socket of a previous tunnel with that ID is closed
fails because the L2TP tunnel context still exists. It is possible to
see this when tunl_delete_race or tunl_query_race apps are run
consecutively like l2tp_ktest does and this results in a test
failure. Fix this by tweaking l2tp_ktest to wait for no tunnels being
present before invoking tunl_delete_race or tunl_query_race.

This behaviour is not observed in kernels newer than v4.17.