tolerate kernels which lack sysctl net.ipv6.conf.all.keep_addr_on_down
sess_datapath.sh sets net.ipv6.conf.all.keep_addr_on_down when it sets
up its test net namespaces. This causes the script to fail on old
kernels that lack the sysctl. But when using those old kernels, we
don't need to set the sysctl anyway, so have the test script just skip
setting the sysctl in that case.