check for kmod refcount leaks
This patch adds a check that l2tp kernel module refcounts are not
leaked when running the tests. By default, the refcounts are compared
before and after a test run. If a problem is detected, refcount leak
checks can be enabled per test. This approach is taken because
refcounts can take a second or so to be released at the end of a test
and the delay per test significantly extends the test runtime.