fix file redirects to kmemleak control files in some environments
If kmemleak is enabled using a Docker/Alpine rootfs, writes to
/sys/kernel/debug/kmemleak result in an error:

echo: write error: Invalid argument

This turns out to be because bash's echo built-in adds a newline when
echo output is redirected to a file and the newline is refused by the
kernel's file handler.

The fix is to use /bin/echo where file redirection is used.