Refactor kernel_session_create to avoid scan-build warning
kernel_session_create had a scan-build warning about dereference of
a null pointer.

The root cause of this was a tricky code path through a goto, meaning
that an assert on entry to the function wasn't actually covering all the
use-cases of the function.

Refactor the code to make it easier to follow and to avoid warning about
the null pointer dereference.