seqnum: make running all tests easier
It's possibly handy to be able to run specific subsets of tests, but by
default it's probably more useful to just run all the seqnum tests.

This is somewhat out of keeping with the other l2tp-ktest tools, which
tend to offer a configurable API and allow a shell script to drive test
suites.  However, sequence number testing is different in nature in that
it's a core function which doesn't vary much depending on L2TP version,
encap types, socket family, etc, etc.