util: expose kernel_session_create_pppox to test applications
kernel_session_create provides a handy turnkey solution for creating a
session instance in the kernel, but it lacks the facility for the caller
to closely control the netlink session configuration when creating
sessions using netlink.

The solution for applications needing to control the netlink
configuration is to use l2tp_netlink directly.  However, when creating
PPP sessions, it's necessary to create pppol2tp sockets to fully
instantate the pseudowire-specific kernel code.

To allow applications to do this, expose the kernel_session_create_pppox
helper function in util.c/h.