seqnum: avoid kernel 'no socket in recv' warning
seqnum uses ppp pseudowires for all its testing since they can be
instantiated for both L2TPv2 and L2TPv3.

Previously, seqnum didn't bother creating pppol2tp sockets for the
sessions it created: it was sufficient to create the session context
using netlink alone since all sequence number handling happens in
l2tp_core.

The downside of this approach is that the kernel's l2tp_ppp code emits a
warning message when receiving a packet which can't be handled because
the ppp infrastructure isn't in place.  Although this doensn't cause a
problem per-se it has potential to confuse the user if they see the
warning.

Tweak seqnum to create pppol2tp sockets when instantiating sessions, in
order to avoid this warning.