netns_datapath: add ac_pppoe datapath test
The test uses four network namespaces, one for a PPPoE client, one
for a LAC, one for an LNS, and one for a router connecting the LAC
and LNS.

At the LAC, an AC_PPPOE L2TP pseudowire is instantiated; at the LNS
a PPP pseudowire is created.

Using pppoe_sess_pkt, we send some random data wrapped in a PPPoE
header, destined for the LAC's eth interface.  The l2tp_ac_pppoe
driver sees the packet, strips the PPPoE header, and forwards the
data payload to the LNS.

The test uses l2tp_core statistics to verify that data has been
transmitted and received over the session.