create netlink socket using libmnl
This is the first step towards replacing libnl3 with libmnl. Create a
genl netlink socket using libmnl.

Old nl-3 code is left in place for now. It will be removed when all nl
code has been converted. Code to be removed is enclosed around a
temporary MNL_PORT define during this work.