port do_nl_send and do_nl_send_recv to libmnl
Define new functions with a suffix so that we can port code over to
libmnl in steps. Drop the timeout on waiting for netlink response
messages since this should never be needed and it is better for the
test script to implement a timeout if one is really necessary.

Leave the existing functions in place until they are no longer
referenced, at which point the temporary suffix can be removed.