fix race on kernel netdev destroy
When l2tp sessions are deleted, the kernel's netdevice is destroyed
asynchronously. Since l2tp-ktest creates sessions with fixed interface
names, this can result in failure to create a session if a netdev
created by a previous test is still being destroyed by the kernel.

To avoid this, have the test script wait for the netdev to be
destroyed where it currently waits for a tunnel to be destroyed in
delay_no_tunnel.