util: squish scan-build warning in _test_init
scan-build warns that we may pass NULL to atoi(3), which could occur if
the env variable OPT_RNG_SEED was unset between two calls to getenv(3).

Avoid this by only calling getenv once and caching the result to pass
to atoi having validated we have a non-NULL string to parse.