netns_datapath: be more explicit when waiting for session interfaces
sess_dataif reports pseudowire interfaces and L2TP IDs to a status file
once the pseudowire is established in the kernel.

For Ethernet and PPP pseudowires, it reports an interface name; however
for PPP/AC there's no interface name to report and hence sess_dataif
logs "?" in the place of an ifname.

Tweak netns_sess_dataif_pair in netns_datapath accordingly to cover the
PPP/AC case.