netns_datapath: use sess_dataif fifo for perf tests
Some of the performance tests create extra tunnels and/or sessions in
the kernel in order to stress the L2TP subsystem data structures used to
track the tunnel/session instances.

To date the code for adding more tunnels and sessions has depended on
iproute2, which limits it to v3 Ethernet pseudowires.

However, now that sess_dataif allows for adding tunnel and session
instances after bringing up the initial link, we can extend the multiple
tunnel/session tests to cover PPP.

As a preparatory step, move the existing tests over to use sess_dataif
in place of iproute2.