sess_dataif: clean up gracefully on signal
The dataplane performance tests which create lots of PPP instances take
a long time to run -- in excess of six minutes, the majority of which is
spent on teardown after the test itself has run.

This appears to be due to PPP instances being closed by default _atexit
style handling somewhere in the process termination code: it's hard to
be certain but it seems there is some racing around PPP subsystem
locking and that adds significant overhead to the shutdown procedure.

We can avoid some of this overhead by gracefully shutting down
sess_dataif on receipt of a signal, because we can schedule things such
that each tunnel socket is closed in turn, hence avoiding lock
contention in the kernel.

In order to do so, the tunnel thread needs to be converted to be
non-blocking (otherwise sess_dataif will hang in cases where a signal is
received while its blocking on socket IO and for whatever reason the
peer isn't going to send any data).

Add a simple state machine for tunnel establishment, and add signal
handlers for SIGTERM/INT/QUIT to allow for graceful shutdown.

This shaves around three minutes off the "lots of PPP instances" perf
test runtime.